Support the new Chromium code search that is served from "http://code.google.com/p/chromium/source/search" but is actually iframing "http://code.google.com/searchframe".
This is done by injecting a content script into the latter that watches for
hashchange events and messages the background page with the frame URL as it
changes.

To avoid having to duplicate this logic into the popup page (which was
duplicating the old tab URL extraction code), the background page now saves the
currently extracted link and popup now gets it from there.